move .class files to a separate folder
modified the script files so that .class files are
moved to proj/classFiles folder. This way commits need not
include .class files, which can be annoying as everytime you
compile they come up as modified. Not including proj/classFiles
in the git repo, so every sync should be followed by a compile.